# main/api-docs/swagger.yaml
swagger: "2.0"
info:
  description: "Ciname API Documentation"
  version: "1.0.0"
  title: "Swagger Cinama"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /users/{id}:
    get:
      tags:
      - "users"
      summary: "Get user by ID"
      description: ""
      operationId: "getUserById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
    delete:
      tags:
      - "users"
      summary: "Delete user by ID"
      description: ""
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The ID that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "User not found"
...
securityDefinitions:
  cinema_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "Modify entity in the system"
      read:pets: "View entity in the system"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      password:
        type: "string"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      active:
        type: "boolean"
      role_id:
        type: "integer"
        format: "int64"
...